requirements.txt
pyTelegramBotAPI==4.31.0
asyncio==4.0.0
aiohttp==3.13.3
python-dotenv==1.2.2
aiohttp-proxy==0.1.2